USB: add the ability to send zero-length packets for USB_SendData
1. You can set flag to (1<<USB_TRANSFERT_FLAG_SEND_ZERO_PKT) to inform the host that the data phase is over in advance.
2. Endpoint 0 control read will automatically add the flag , while other endpoints need to be added manually as needed.
3. Not all endpoint 0 control-read-cmd will automatically add this flag, for example, GET DEVICE DESCRIPTOR.
3. This flag only takes effect in certain scenarios:
(1) The size of the data is going to be sent is an exactly multiple of mps.
(2) The size of the data is going to be sent  is less than the size that the host wants to get. Or the host does not know the total length of the device to be sent in advance.